+++ es-server/src/main/java/io/joshworks/eventry/server/cluster/messages/CreateStream.java
@@ -2,14 +2,14 @@ package io.joshworks.eventry.server.cluster.messages;
 
 import java.util.Map;
 
-public class CreateStream  {
+public class CreateStream {
     public final String stream;
     public final int maxCount;
-    public final  long maxAge;
-    public final  Map<String, Integer> acl;
-    public final  Map<String, String> metadata;
+    public final int maxAge;
+    public final Map<String, Integer> acl;
+    public final Map<String, String> metadata;
 
-    public CreateStream(String stream, int maxCount, long maxAge, Map<String, Integer> acl, Map<String, String> metadata) {
+    public CreateStream(String stream, int maxCount, int maxAge, Map<String, Integer> acl, Map<String, String> metadata) {
         this.stream = stream;
         this.maxCount = maxCount;
         this.maxAge = maxAge;
